♻️ Extraced NameInput.vue from NwkCreateDialog.vue
