+++ praktikumsplaner-frontend/frontend/src/components/nachwuchskraefte/NwkCreateDialog.vue
@@ -23,22 +23,18 @@
                         <v-row>
                             <v-col>
                                 <v-list-item>
-                                    <v-text-field
-                                        v-model="nwk.vorname"
-                                        label="Vorname"
-                                        :rules="nameRule"
-                                        outlined
-                                    ></v-text-field>
+                                    <NameInput
+                                        :nwk="nwk"
+                                        type="nachname"
+                                    ></NameInput>
                                 </v-list-item>
                             </v-col>
                             <v-col>
                                 <v-list-item>
-                                    <v-text-field
-                                        v-model="nwk.nachname"
-                                        label="Nachname"
-                                        :rules="nameRule"
-                                        outlined
-                                    ></v-text-field>
+                                    <NameInput
+                                        :nwk="nwk"
+                                        type="vorname"
+                                    ></NameInput>
                                 </v-list-item>
                             </v-col>
                         </v-row>
@@ -105,22 +101,12 @@ import NwkCreate from "@/types/NwkCreate";
 import FetchUtils from "@/api/FetchUtils";
 import VorlesungstageSelector from "@/components/nachwuchskraefte/VorlesungstageSelect.vue";
 import StudienrichtungOrAusbildungsrichtungSelect from "@/components/common/StudienrichtungOrAusbildungsrichtungSelect.vue";
+import NameInput from "@/components/common/NameInput.vue";
 
 const visible = ref<boolean>(false);
 const loading = ref<boolean>(false);
 const form = ref<HTMLFormElement>();
 const validationRules = useRules();
-const nameRule = [
-    validationRules.minLengthRule(
-        2,
-        "Der Name muss mindestens 2 Zeichen lang sein."
-    ),
-    validationRules.maxLengthRule(
-        255,
-        "Der Name darf maximal 255 Zeichen lang sein."
-    ),
-    validationRules.notEmptyRule("Der Name darf nicht leer sein."),
-];
 const jahrgangRule = [
     validationRules.notEmptyRule("Der Jahrgang darf nicht leer sein."),
     validationRules.regexRule(
